renamed angular environments folder correctly
