+++ indexcards/src/main/resources/application.properties
@@ -1,7 +1,7 @@
 spring.jpa.hibernate.ddl-auto=update
 spring.datasource.url=jdbc:mysql://${MYSQL_HOST:localhost}:3306/indexcardsdb
 spring.datasource.username=root
-spring.datasource.password=12345678
+spring.datasource.password=1234
 spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
 spring.main.allow-bean-definition-overriding=true
